cdk/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
attrs==22.2.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.49
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.1
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v5==2.0.38
    # via aws-cdk-lib
aws-cdk-aws-lambda-python-alpha==2.61.1a0
    # via -r requirements.in
aws-cdk-lib==2.61.1
    # via
    #   -r requirements.in
    #   aws-cdk-aws-lambda-python-alpha
boto3==1.26.54
    # via -r requirements.in
botocore==1.29.54
    # via
    #   boto3
    #   s3transfer
cattrs==22.2.0
    # via jsii
click==8.1.3
    # via mcstatus
constructs==10.1.227
    # via
    #   -r requirements.in
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
dnspython==1.15.0
    # via dnspython3
dnspython3==1.15.0
    # via mcstatus
exceptiongroup==1.1.0
    # via cattrs
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsii==1.73.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   constructs
marshmallow==3.19.0
    # via -r requirements.in
mcstatus==2.3.0
    # via -r requirements.in
packaging==23.0
    # via marshmallow
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   constructs
    #   jsii
python-dateutil==2.8.2
    # via
    #   botocore
    #   jsii
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via
    #   mcstatus
    #   python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   constructs
    #   jsii
typing-extensions==4.4.0
    # via jsii
urllib3==1.26.14
    # via botocore
